requirements.txt
asgiref==3.3.1
dj-static==0.0.6
Django==3.1.5
django-simple-history==2.12.0
djangorestframework==3.12.2
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
gunicorn==20.0.4
virtualenv==20.3.1
virtualenv==20.3.1
whitenoise==5.2.0
wrapt==1.12.1
